requirements.txt
crewai>=0.28.8
python-dotenv>=1.0.0
openai>=1.12.0
notion-client>=2.2.0
slack-sdk>=3.27.0
requests>=2.31.0
pytest>=7.4.0